mcp-servers/data-mcp/requirements.txt
mcp[cli]>=1.0.0
asyncpg>=0.29.0
httpx>=0.27.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0